# src/generated/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Transfeera API",
    "version": "0.0.1",
    "description": "An API for simple CRUD Recipients."
  },
  "host": "0.0.0.0:3000",
  "paths": {
    "/api/recipients/bulk-delete": {
      "post": {
        "tags": [
          "Recipient"
        ],
        "description": "Delete many `Recipient` by its `ids`",
        "requestBody": {
          "required": true,
          "description": "A list of `Recipient`'s ids.",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ids": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                },
                "required": [
                  "ids"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "deletion success.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "total": {
                      "type": "integer",
                      "description": "The total of `Recipient` deleteds."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "invalid json body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequest"
                }
              }
            }
          },
          "500": {
            "description": "Internal Error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      }
    },
    "/api/recipients": {
      "post": {
        "tags": [
          "Recipient"
        ],
        "description": "Create a new `Recipient`",
        "requestBody": {
          "required": true,
          "description": "The `CreateUpdateRecipientPayload` data.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEditRecipientPayload"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "deletion success.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Recipient"
                }
              }
            }
          },
          "400": {
            "description": "invalid json body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequest"
                }
              }
            }
          },
          "500": {
            "description": "Internal Error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Recipient"
        ],
        "description": "Returns a list of `Recipient`",
        "parameters": [
          {
            "in": "query",
            "name": "itemsPerPage",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "The quantity of items to be returned per page."
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "VALIDADO",
                "RASCUNHO"
              ]
            },
            "description": "The `Status` of the recipient."
          },
          {
            "in": "query",
            "name": "name",
            "schema": {
              "type": "string"
            },
            "description": "The name of the recipient."
          },
          {
            "in": "query",
            "name": "key_type",
            "schema": {
              "type": "string",
              "enum": [
                "CPF",
                "CNPJ",
                "EMAIL",
                "TELEFONE",
                "CHAVE_ALEATORIA"
              ]
            },
            "description": "The Pix's Key types of the recipient."
          },
          {
            "in": "query",
            "name": "key_value",
            "schema": {
              "type": "integer"
            },
            "description": "The Pix's key value of the recipient."
          }
        ],
        "responses": {
          "200": {
            "description": "list success.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "total": {
                      "type": "integer",
                      "description": "The total of `Recipient` data returned."
                    },
                    "data": {
                      "description": "The list of `Recipient` data.",
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Recipient"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "list success.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequest"
                }
              }
            }
          },
          "500": {
            "description": "Internal Error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      }
    },
    "/api/recipients/{recipientID}": {
      "delete": {
        "tags": [
          "Recipient"
        ],
        "description": "Delete a `Recipient` by its `id`",
        "parameters": [
          {
            "in": "path",
            "name": "recipientID",
            "schema": {
              "type": "string",
              "required": true
            },
            "description": "The `Recipient` identifier."
          }
        ],
        "responses": {
          "200": {
            "description": "deletion success."
          },
          "400": {
            "description": "invalid recipient value.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequest"
                }
              }
            }
          },
          "500": {
            "description": "Internal Error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Recipient"
        ],
        "description": "Update a `Recipient`",
        "parameters": [
          {
            "name": "recipientId",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "description": "The `Recipient` identifier.",
            "required": true
          }
        ],
        "requestBody": {
          "required": true,
          "description": "The `CreateUpdateRecipientPayload` data.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEditRecipientPayload"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "update success.",
            "content": null
          },
          "400": {
            "description": "invalid json body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequest"
                }
              }
            }
          },
          "500": {
            "description": "Internal Error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "KeyTypes": {
      "type": "string",
      "enum": [
        "CPF",
        "CNPJ",
        "EMAIL",
        "TELEFONE",
        "CHAVE_ALEATORIA"
      ]
    },
    "Status": {
      "type": "string",
      "enum": [
        "VALIDADO",
        "RASCUNHO"
      ]
    }
  },
  "components": {
    "schemas": {
      "InternalServerError": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "description": "HTTP Status Code.",
            "default": 500
          },
          "error": {
            "type": "string",
            "description": "HTTP Message Code.",
            "default": "Internal Server Error"
          },
          "message": {
            "type": "string",
            "description": "A message describing the reason for the error."
          }
        }
      },
      "BadRequest": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "description": "HTTP Status Code.",
            "default": 400
          },
          "error": {
            "type": "string",
            "description": "HTTP Message Code.",
            "default": "Bad Request"
          },
          "message": {
            "type": "string",
            "description": "A message describing the reason for the error."
          }
        }
      },
      "Recipient": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "cpf_cnpj": {
            "type": "string"
          },
          "key_type": {
            "type": "string",
            "enum": [
              "CPF",
              "CNPJ",
              "EMAIL",
              "TELEFONE",
              "CHAVE_ALEATORIA"
            ]
          },
          "key_value": {
            "type": "string"
          },
          "bank": {
            "type": "string"
          },
          "agency": {
            "type": "string"
          },
          "account": {
            "type": "string"
          },
          "account_type": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "VALIDADO",
              "RASCUNHO"
            ]
          }
        },
        "required": [
          "id",
          "name",
          "cpf_cnpj",
          "status"
        ]
      },
      "CreateEditRecipientPayload": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "cpf_cnpj": {
            "type": "string"
          },
          "key_type": {
            "type": "string",
            "enum": [
              "CPF",
              "CNPJ",
              "EMAIL",
              "TELEFONE",
              "CHAVE_ALEATORIA"
            ]
          },
          "key_value": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "cpf_cnpj",
          "key_type",
          "key_value"
        ]
      },
      "recipientID": {
        "type": "string",
        "format": "uuid"
      }
    }
  },
  "tags": [
    {
      "name": "Recipient",
      "description": "Hello world end point"
    }
  ]
}